ポモドーロ機能のリファクタリング: utils.jsを削除し、機能をpomodoroCalculator.js、pomodoroMessage.js、pomodoroState.js、pomodoroVoice.jsに分割。依存関係を整理し、コードの可読性と保守性を向上。
