+++ lib/pomodoro/main.js
@@ -4,7 +4,18 @@
 
 const path = require('path');
 const dirname = require('../defineDirname.js');
-const pomodoroUtils = require(path.join(dirname, 'lib/pomodoro/utils.js'));
+const pomodoroStateUtils = require(
+	path.join(dirname, 'lib/pomodoro/utils/pomodoroState.js'),
+);
+const pomodoroCalculatorUtils = require(
+	path.join(dirname, 'lib/pomodoro/utils/pomodoroCalculator.js'),
+);
+const pomodoroMessageUtils = require(
+	path.join(dirname, 'lib/pomodoro/utils/pomodoroMessage.js'),
+);
+const pomodoroVoiceUtils = require(
+	path.join(dirname, 'lib/pomodoro/utils/pomodoroVoice.js'),
+);
 const timerController = require(
 	path.join(dirname, 'lib/pomodoro/timerController.js'),
 );
@@ -13,7 +24,10 @@ const { EmbedBuilder, MessageFlags } = require('discord.js');
 
 async function start(client, interaction, options = {}) {
 	const guildId = interaction.guildId;
-	const pomodoroState = pomodoroUtils.getPomodoroState(client, guildId);
+	const pomodoroState = await pomodoroStateUtils.getPomodoroState(
+		client,
+		guildId,
+	);
 
 	if (pomodoroState.running) {
 		await interaction.reply('❌ すでにポモドーロタイマーが実行されています。');
@@ -99,9 +113,9 @@ async function start(client, interaction, options = {}) {
 			pomodoroState.part.cycle % timesUntilLongBreak === 0
 				? 'longBreak'
 				: 'break';
-		await pomodoroUtils.sendPomodoroStatus(interaction, pomodoroState);
+		await pomodoroMessageUtils.sendPomodoroStatus(interaction, pomodoroState);
 		if (voiceNotification) {
-			await pomodoroUtils.notifyVoice(
+			await pomodoroVoiceUtils.notifyVoice(
 				client,
 				interaction.guild.id,
 				pomodoroState,
@@ -118,9 +132,9 @@ async function start(client, interaction, options = {}) {
 			pomodoroState.part.endTimestamp = longBreakEndTime;
 			pomodoroState.part.type = 'longBreak';
 			pomodoroState.part.nextType = 'work';
-			await pomodoroUtils.sendPomodoroStatus(interaction, pomodoroState);
+			await pomodoroMessageUtils.sendPomodoroStatus(interaction, pomodoroState);
 			if (voiceNotification) {
-				await pomodoroUtils.notifyVoice(
+				await pomodoroVoiceUtils.notifyVoice(
 					client,
 					interaction.guild.id,
 					pomodoroState,
@@ -133,9 +147,9 @@ async function start(client, interaction, options = {}) {
 			pomodoroState.part.endTimestamp = breakEndTime;
 			pomodoroState.part.type = 'break';
 			pomodoroState.part.nextType = 'work';
-			await pomodoroUtils.sendPomodoroStatus(interaction, pomodoroState);
+			await pomodoroMessageUtils.sendPomodoroStatus(interaction, pomodoroState);
 			if (voiceNotification) {
-				await pomodoroUtils.notifyVoice(
+				await pomodoroVoiceUtils.notifyVoice(
 					client,
 					interaction.guild.id,
 					pomodoroState,
@@ -177,7 +191,7 @@ async function start(client, interaction, options = {}) {
 		}
 
 		if (voiceNotification && pomodoroState.part.type !== 'cancelled') {
-			await pomodoroUtils.notifyVoice(
+			await pomodoroVoiceUtils.notifyVoice(
 				client,
 				interaction.guild.id,
 				pomodoroState,
@@ -185,14 +199,17 @@ async function start(client, interaction, options = {}) {
 			);
 		}
 
-		await pomodoroUtils.clearPomodoroState(client, guildId);
+		await pomodoroStateUtils.clearPomodoroState(client, guildId);
 	}
 }
 
 // 一時停止
 async function pause(client, interaction) {
 	const guildId = interaction.guild.id;
-	const pomodoroState = pomodoroUtils.getPomodoroState(client, guildId);
+	const pomodoroState = await pomodoroStateUtils.getPomodoroState(
+		client,
+		guildId,
+	);
 
 	// ポモドーロタイマーが実行中でない、または既に一時停止中の場合はエラー
 	if (!pomodoroState.running || pomodoroState.paused) {
@@ -207,7 +224,7 @@ async function pause(client, interaction) {
 
 	// 残り時間を凍結してタイマーを一時停止
 	pomodoroState.part.pausedRemainingSeconds =
-		pomodoroUtils.calculateRemainingSeconds(pomodoroState);
+		pomodoroCalculatorUtils.calculateRemainingSeconds(pomodoroState);
 	pomodoroState.part.endTimestamp = null;
 	pomodoroState.paused = true;
 	timerController.pause(pomodoroState);
@@ -220,7 +237,7 @@ async function pause(client, interaction) {
 		const currentDescription = message.embeds[0]?.description ?? '';
 		const baseDescription =
 			currentDescription.split('次のパート: ')[0] ?? currentDescription;
-		const updatedEmbed = pomodoroUtils.updateEmbedDescription(
+		const updatedEmbed = pomodoroMessageUtils.updateEmbedDescription(
 			message.embeds[0],
 			`${baseDescription}次のパート: \`一時停止中\``,
 		);
@@ -238,7 +255,10 @@ async function pause(client, interaction) {
 // 再開
 async function resume(client, interaction) {
 	const guildId = interaction.guild.id;
-	const pomodoroState = pomodoroUtils.getPomodoroState(client, guildId);
+	const pomodoroState = await pomodoroStateUtils.getPomodoroState(
+		client,
+		guildId,
+	);
 
 	// 一時停止中でない場合はエラー
 	if (!pomodoroState.paused) {
@@ -273,7 +293,7 @@ async function resume(client, interaction) {
 		pomodoroState.message.id,
 	);
 	if (message) {
-		const updatedEmbed = pomodoroUtils.updateEmbedDescription(
+		const updatedEmbed = pomodoroMessageUtils.updateEmbedDescription(
 			message.embeds[0],
 			(message.embeds[0]?.description ?? '').replace(
 				/<t:(\d+):R>/g,
@@ -294,7 +314,10 @@ async function resume(client, interaction) {
 // 状況確認
 async function status(client, interaction) {
 	const guildId = interaction.guildId;
-	const pomodoroState = pomodoroUtils.getPomodoroState(client, guildId);
+	const pomodoroState = await pomodoroStateUtils.getPomodoroState(
+		client,
+		guildId,
+	);
 
 	if (!pomodoroState.running && !pomodoroState.paused) {
 		await interaction.reply('❌ ポモドーロタイマーは動作していません。');
@@ -303,7 +326,7 @@ async function status(client, interaction) {
 
 	// リアルタイムで残り時間を計算
 	const currentRemainingSeconds =
-		pomodoroUtils.calculateRemainingSeconds(pomodoroState);
+		await pomodoroCalculatorUtils.calculateRemainingSeconds(pomodoroState);
 	const remaining =
 		typeof currentRemainingSeconds === 'number'
 			? `${String(Math.floor(currentRemainingSeconds / 60)).padStart(
@@ -321,7 +344,10 @@ async function status(client, interaction) {
 // 終了
 async function stop(client, interaction) {
 	const guildId = interaction.guild.id;
-	const pomodoroState = pomodoroUtils.getPomodoroState(client, guildId);
+	const pomodoroState = await pomodoroStateUtils.getPomodoroState(
+		client,
+		guildId,
+	);
 	await interaction.deferReply({ flags: MessageFlags.Ephemeral });
 
 	if (!pomodoroState.running && !pomodoroState.paused) {
