+++ lib/pomodoro/timerController.js
@@ -1,6 +1,11 @@
 const path = require('path');
 const dirname = require('../defineDirname.js');
-const pomodoroUtils = require(path.join(dirname, 'lib/pomodoro/utils.js'));
+const pomodoroCalculatorUtils = require(
+	path.join(dirname, 'lib/pomodoro/utils/pomodoroCalculator.js'),
+);
+const pomodoroMessageUtils = require(
+	path.join(dirname, 'lib/pomodoro/utils/pomodoroMessage.js'),
+);
 
 // BOT全体でのポモドーロタイマーの自動更新のレートを監視するための変数
 let totalAutoUpdateCount = 0;
@@ -34,7 +39,7 @@ function startInterval(client, interaction, pomodoroState) {
 
 		// 時間が終了したかチェック
 		const remainingSeconds =
-			pomodoroUtils.calculateRemainingSeconds(pomodoroState);
+			await pomodoroCalculatorUtils.calculateRemainingSeconds(pomodoroState);
 		if (remainingSeconds <= 0) {
 			// パートの時間が終了した場合はタイマーをクリアしてresolve
 			stop(pomodoroState, 'completed');
@@ -85,10 +90,10 @@ function startInterval(client, interaction, pomodoroState) {
 		}
 		// 現在のパートの割合を計算
 		const statusPercentage =
-			pomodoroUtils.calculateStatusPercentage(pomodoroState);
+			await pomodoroCalculatorUtils.calculateStatusPercentage(pomodoroState);
 		// サーバーへの導入数次第で、絵文字の割合を計算
 		const statusEmojis =
-			pomodoroUtils.generatePomodoroStatusEmojis(statusPercentage);
+			await pomodoroMessageUtils.generatePomodoroStatusEmojis(statusPercentage);
 		let shouldUpdate = false;
 		const doneCountDiff = Math.abs(
 			statusEmojis.doneCount - pomodoroState.message.lastProgress,
@@ -126,7 +131,10 @@ function startInterval(client, interaction, pomodoroState) {
 		if (pomodoroState.running && shouldUpdate && totalAutoUpdateCount < 10) {
 			// ステータスを更新する
 			try {
-				await pomodoroUtils.sendPomodoroStatus(interaction, pomodoroState);
+				await pomodoroMessageUtils.sendPomodoroStatus(
+					interaction,
+					pomodoroState,
+				);
 			} catch (err) {
 				// 送信失敗は無視
 				void err;
